pyproject.toml
[project]
name = "homelab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "attrs>=25.3.0",
    "fire>=0.7.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "boto3-stubs>=1.40.39",
    "flake8>=7.3.0",
    "mypy>=1.18.2",
    "pdbpp>=0.11.7",
    "pycodestyle>=2.14.0",
    "pycompile>=0.2.3",
    "pyflakes>=3.4.0",
    "pylint>=3.3.8",
    "pytest>=8.4.2",
    "pytest-html>=4.1.1",
    "python-lsp-server>=1.13.1",
    "ruff>=0.13.2",
    "ruff-lsp>=0.0.62",
    "semgrep>=1.138.0",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.uv]
package = true
reinstall-package = ["homelab"]
